Lock Mephisto to Rails 2.0.5 gem
Mephisto is only compatible with Rails 2.0.  The latest security fix for
the 2.0 series is 2.0.5.  But many Rails developers will have Rails 2.2
installed on their systems.  So this patch locks Mephisto to an appropriate
version of the Rails gem, allowing it to run on systems with Rails 2.2
installed.

This patch also includes the changes applied by 'rake rails:update', minus
the Prototype JavaScript files which had been removed from the original
Mephisto.