Fix more problems caused by protect_action
The :approve and :unapprove actions are destructive, and should
therefore only be called with :post, and never with :get.

The theme :import action is a bit more interesting.  It can be called
with :get, which should return an import form, so it doesn't belong on
the list of protected actions.