Fix TRUNCATE TABLE errors in in-tree model_stubbing
As discovered by Aubrey Holland, model_stubbing relies on the
non-portable TRUNCATE TABLE command:

  http://groups.google.com/group/MephistoBlog/msg/7da91b7410271ff2

Applying Aubrey's fix allows 'rake spec' to run against SQLite3.  Note
that there are still a good number of spec failures which will need to
be addressed.

It's a little bit dodgy to patch the in-tree copy of model_stubbing,
but doing so will suffice for now.