Properly reload article after commenting disallowed
When running the integration tests, I received the following warning:

  test_should_not_cache_comment_post_on_article_with_closed_comments
  (CachingTest): .../mephisto/vendor/plugins/liquid/lib/liquid/drop.rb:33:
  warning: Object#id will be deprecated; use Object#object_id

This was caused by a stale comment object attached to @article, which
caused "comment.id" in a Liquid template to issue a warning.  The stale
comment was caused by attempting to add a new comment to @article and
encountering an error because comments were closed.

The fix is to reload the article before rendering the template, causing
the stale comment to be dropped.