Do not short-circuit 'render :update' block with return
Using 'return' in render :update has stopped working as of Ruby 1.8.7.
(It causes Rails to attempt to render search.html.erb instead of
returning the newly-rendered JavaScript.)  I hope this is just an
isolated problem, and not part of something systemic.

The unit tests are now all passing under Ruby 1.8.7.