Don't accidentally save article when user searches for assets
When the user clicks on the search tab, types some text, and hits
enter, it triggers a form submission for the main article form.  This is
very annoying.

So now we intercept the keypress event on this field, stop it from
propagating, and instead trigger the same logic as the "Search" button.
I found the following helpful for getting started:

  http://ne0phyte.com/blog/2008/09/02/javascript-keypress-event/