Fix another test failure that only occurs at certain times
This test case would fail for about a day or so near the third of the
month, when the articles in archive/YEAR/MONTH weren't what it expected,
because the article dates were assigned dynamically.

We should think about switching more of these tests over to machinist,
actually.  The fixtures are fairly awkward, especially with all the
dynamic dates.