replace use of '<<' operator with '+=' operator in find_child_sections and find_decendant_sections in SiteDrop
We found that mutiple calls to the 'child_sections' drop filter was modifying the value of the path
attribute inside of SectionDrops.  The SiteDrop methods find_child_sections and find_decentdent_sections
used the '<<' operator to append a '/' charecter to the path.  This had the side effect of modifying
the original string in memory.  The "+=" while slower, uses a copy of the string passed as parameter.

The gave us the behavior we would expect.