Clean up lots of deprecation warnings
These fall into two categories:
  - Render statements that now need an explicit :object
  - Truncate statements that need keywords arguments

The are still warnings about ActiveRecord::Errors.default_error_messages
that appear in test:functionals, but that only gets called from
attachment_fu, and not from any obvious place in app, so I'm going to
punt on it for a little bit.

This leaves us with 2 failures, 6 errors and 2 warnings.  We're getting
really close.