Start using config.gem to manage gems
These gems can all be moved easily without breaking our unit tests.  Use
'rake gems', etc., to list and manage gems declared with config.gem.

We update a couple of gems here, most notable rubyzip.  We don't update
to the newest version of RedCloth, because Mephisto seemed to want a
specific version before, and the newest version includes a native
extension that we don't want to deal with right now.