Fix sections/create.rjs to show feedback to user
[Notes from Eric.

To reproduce, create a new section.  What happens: The section gets
created, but there is no user feedback.  What happens with this patch
applied: The user gets a flash message and the new section appears in
the sidebar.

However, we still don't clear the "New Section" form, which is a bit
confusing.  But this is definitely a net improvement over what we
have now.  Many thanks, James!]