Move observer setup to environment.rb
I'm working on a fix for the problem reported here:

  http://github.com/emk/mephisto/commit/dadfc5765d738b902c5cef8f293e865e2c536239#comment_9394

Basically, 'rake db:create' is trying to load observers, and this in
turn tries to load a whole bunch of the app before the database exists.

This patch doesn't fix the problem, but it does make our initialization
a bit more modern.