Security: Replace white_list with Rails 2.2 sanitizer
The Rails 2.2 santizer is an enhanced version of Rick's original
white_list plugin, so let's upgrade and get the latest fixes.

Note that Mephisto had separate rules for sanitizing comments and
non-comments in Atom feeds.  This difference was introduced in commit
88df87e3a1cb8474fa479d855035ab4d2ca2351e.  Unfortunately, I'm not able
to track down any information on the problem being fixed here.  Since we
already add half of the tags in question to the whitelist, I've decided
to just treat all sanitized Atom feed content the same.  Please let me
know if this breaks anything.