Make 'rake gems' more robust when gems are missing
When 'config.gem' fails to load a gem, Rails appears to go ahead and
try to load routes anyway.  But in this scenario, the engines plugin
doesn't get loaded either, and so we can't use it to load plugin routes.

This patch attempts to make 'rake gems' a bit more robust in the face
of these kinds of errors.  Unforunately, it may also cause mysterious
failures of plugin routing if the engines plugin goes unloaded for
other reasons, so I've included a warning.  This may not be entirely
adequate, but it's better than breaking 'rake gems'.