Fix theme controller bugs
J.C. Zhu reported an InvalidAuthenticityToken error when trying to
change themes.  This was one of several bugs in the theme switching
code left over from the Rails 2.2 porting process and the security
audit.

We needed to convince Ajax.Request to use a GET request when displaying
the theme tools, and we needed to properly escape some text in the
_tools template.

I've also added a test case to make sure we actually render the _tools
view successfully.