Modernize rjs: admin/articles
I'm going to try to rename all the *.rjs files to *.js.rjs.  This is
trickier than the other renamings, because our unit test coverage isn't
perfect, and I'm trying to test everything by hand when possible.  So
I'm going to do this one directory at a time.

Other changes:
  - The live_preview and _preview stuff wasn't being used.
  - We didn't have a test case for the 'label' action.