Begin updating to latest Prototype
Our version of Prototype is really old, and no longer compatible with
current versions of RJS.  This breaks quite a bit of our admin interace
in subtle ways.

The patch upgrades our bundled version of Prototype to the version
included with Rails 2.2.2.  We also remove some obsolete bits of
application.js and make a few minor fixes so that it loads without
errors.

I've only tested the JavaScript support on the articles page and the
assets page.  Quite a bit of formerly broken stuff now seems to work,
including updates to the list of attached files and updates to the
contents of the bucket.  But there's still more to test and fix.