Write login integration tests using Webrat
Why do we need integration tests? We've been suffering a lot of
regressions in the Mephisto UI, because our functional tests don't reach
high enough up towards the browser, and whole classes of bugs manage to
slip through.

What is Webrat? Webrat is a "browser simulator" written in Ruby.  It
generates a DOM and allows us to fill in forms as though an actual user
were interacting with the site.

Why Webrat, and not Selenium, Watir, etc?  Webrat is recommended by the
Cucumber project as the default way to write user stories; it's very
fast; and it has a reasonable API.  Plus, Webrat actively maintained,
and very recent versions of Webrat can be used as a front-end to
Selenium.

Why Rails integration tests, and not Cucumber stories?  Since the people
contributing to Mephisto will largely be programmers, I decided to write
integration tests using a Ruby-based DSL.  Cucumber stories look really
interesting, but with no actual clients in the loop, the text-based format
is slightly less useful and has a steeper learning curve for programmers.

Since we're switching to a new integration testing tool, I moved a bunch of
code out of test_helper.rb and put it into our only existing integration
test, caching_test.rb.  I also switched blueprints.rb to set up user
passwords using 'password' and 'password_confirm' (instead of crypted and
salted values) to make it easier for tests to override